+++ thrustTest.cu
@@ -29,13 +29,13 @@ int main(){
 	size_t m = (NODES*NODES)/2 - NODES/2;
 
 	curandGenerator_t gen, gen1;
-	float *devData, *hostData, *devData1, *hostAdjacent;
+	double *devData, *hostData, *devData1, *hostAdjacent;
 
-	hostData = (float *)calloc(n, sizeof(float));
-	hostAdjacent = (float *)calloc(m, sizeof(float));
+	hostData = (double *)calloc(n, sizeof(double));
+	hostAdjacent = (double *)calloc(m, sizeof(double));
 
-	CUDA_CALL(cudaMalloc((void **)&devData, n*sizeof(float)));
-	CUDA_CALL(cudaMalloc((void **)&devData1, m*sizeof(float)));
+	CUDA_CALL(cudaMalloc((void **)&devData, n*sizeof(double)));
+	CUDA_CALL(cudaMalloc((void **)&devData1, m*sizeof(double)));
 
 	CURAND_CALL(curandCreateGenerator(&gen, CURAND_RNG_PSEUDO_DEFAULT));
 	CURAND_CALL(curandCreateGenerator(&gen1, CURAND_RNG_PSEUDO_DEFAULT));
@@ -46,8 +46,8 @@ int main(){
 	CURAND_CALL(curandGenerateUniformfloat(gen, devData, n));
 	CURAND_CALL(curandGenerateUniformfloat(gen1, devData1, m));
 
-	CUDA_CALL(cudaMemcpy(hostData, devData, n * sizeof(float), cudaMemcpyDeviceToHost));
-	CUDA_CALL(cudaMemcpy(hostAdjacent, devData1, m * sizeof(float), cudaMemcpyDeviceToHost));
+	CUDA_CALL(cudaMemcpy(hostData, devData, n * sizeof(double), cudaMemcpyDeviceToHost));
+	CUDA_CALL(cudaMemcpy(hostAdjacent, devData1, m * sizeof(double), cudaMemcpyDeviceToHost));
 
 	for(i = 0; i < 10; i++){
 		for(size_t j = 0; j < 10; j++){
@@ -62,17 +62,17 @@ int main(){
 	CURAND_CALL(curandDestroyGenerator(gen1));
 	CUDA_CALL(cudaFree(devData1));
 
-	thrust::host_vector<float> H(hostData, hostData + NODES);
+	thrust::host_vector<double> H(hostData, hostData + NODES);
 
 	std::cout << "H has size " << H.size() << std::endl;
 
-	thrust::device_vector<float> D = H;
+	thrust::device_vector<double> D = H;
 
-//	float* maxResult = thrust::max_element(thrust::host, D.begin(), D.end());
+//	double* maxResult = thrust::max_element(thrust::host, D.begin(), D.end());
 
-	thrust::detail::normal_iterator<thrust::device_ptr<float> > maxResult = thrust::max_element(D.begin(), D.end());
+	thrust::detail::normal_iterator<thrust::device_ptr<double> > maxResult = thrust::max_element(D.begin(), D.end());
 
-	//thrust::host_vector<float> h_maxResult = maxResult;
+	//thrust::host_vector<double> h_maxResult = maxResult;
 
 	std::cout << "maxResult is " << *maxResult << std::endl;
